pi_brain/requirements.txt
pyserial
google-generativeai
Pillow
pypdfium2
python-dotenv